+++ packages/kap-server/test/skills.test.ts
@@ -21,7 +21,7 @@
  * `routes/skills.ts`, which must match the session listing for the same cwd.
  */
 
-import { mkdir, mkdtemp, rm, writeFile } from 'node:fs/promises';
+import { mkdir, mkdtemp, readFile, rm, writeFile } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 
@@ -262,6 +262,74 @@ describe('server-v2 /api/v1 skills', () => {
       );
       expect(body.code).toBe(40001);
       expect(body.msg).toMatch(/unsupported action/);
+    });
+
+    it('carries a file attachment into the activation message', async () => {
+      const id = await createSession();
+      await createMainAgent(id);
+
+      const noteBytes = Buffer.from('hello from the attachment');
+      const form = new FormData();
+      form.set('file', new Blob([noteBytes], { type: 'text/plain' }), 'note.txt');
+      const uploadRes = await fetch(`${base}/api/v1/files`, {
+        method: 'POST',
+        headers: authHeaders(server as RunningServer),
+        body: form,
+      } as never);
+      const uploaded = (await uploadRes.json()) as Envelope<{ id: string; size: number }>;
+      expect(uploaded.code).toBe(0);
+
+      const { body } = await postJson<{ activated: boolean; skill_name: string }>(
+        `/api/v1/sessions/${id}/skills/update-config:activate`,
+        {
+          args: '--help',
+          attachments: [
+            {
+              type: 'file',
+              file_id: uploaded.data.id,
+              name: 'note.txt',
+              media_type: 'text/plain',
+              size: noteBytes.length,
+            },
+          ],
+        },
+      );
+      expect(body.code).toBe(0);
+      expect(body.data).toEqual({ activated: true, skill_name: 'update-config' });
+
+      // The activation's user message carries the rendered skill prompt
+      // followed by the materialized attachment's path notice — the same
+      // pipeline a prompt submission runs through.
+      const messages = await getJson<{
+        items: Array<{ role: string; content: Array<{ type: string; text?: string }> }>;
+      }>(`/api/v1/sessions/${id}/messages`);
+      const userMsg = messages.body.data.items.find((m) => m.role === 'user');
+      expect(userMsg).toBeDefined();
+      expect(userMsg!.content[0]?.type).toBe('text');
+      expect(userMsg!.content[0]?.text).toContain('User activated the skill "update-config"');
+      const notice = userMsg!.content[1];
+      expect(notice?.type).toBe('text');
+      expect(notice?.text).toContain('Attached file "note.txt"');
+      expect(notice?.text).toContain(`${noteBytes.length} bytes`);
+      const attachedPath = /bytes\): (.+) — open it with the Read tool$/.exec(notice?.text ?? '')?.[1];
+      expect(attachedPath).toBeDefined();
+      expect(attachedPath).toContain('/attachments/');
+      expect(await readFile(attachedPath!)).toEqual(noteBytes);
+    });
+
+    it('rejects an activation with a stale attachment file_id (40407)', async () => {
+      const id = await createSession();
+      await createMainAgent(id);
+
+      const { body } = await postJson<null>(
+        `/api/v1/sessions/${id}/skills/update-config:activate`,
+        {
+          attachments: [
+            { type: 'file', file_id: 'f_does_not_exist', name: 'x.txt', media_type: 'text/plain', size: 1 },
+          ],
+        },
+      );
+      expect(body.code).toBe(40407);
     });
   });
